Plugins: categorize indirect_system_calls as a malware plugin
